Fixes the-teacher/the_comments#26 - ActiveSupport::Concern::MultipleIncludedBlocks
