+++ app/models/concerns/the_comments/comment.rb
